backend/requirements.txt
fastapi>=0.115,<0.120
uvicorn[standard]>=0.32,<0.40
motor>=3.6
pydantic>=2.10
pydantic-settings>=2.6
python-dotenv>=1.0
httpx>=0.27
chromadb>=0.5.20
cloudinary>=1.41
PyJWT>=2.9
python-jose[cryptography]>=3.3
email-validator>=2.2
